MAINT: Move plot logic from restoration to plotting
