+++ src/spectral_recovery/restoration.py
@@ -18,13 +18,8 @@ from inspect import signature
 import xarray as xr
 import geopandas as gpd
 import pandas as pd
-import matplotlib.pyplot as plt
-import seaborn as sns
 
 from pandas import Index as pdIndex
-from matplotlib.lines import Line2D
-from matplotlib.patches import Patch
-from matplotlib.legend_handler import HandlerPatch
 
 from spectral_recovery.targets import MedianTarget, expected_signature
 from spectral_recovery.timeseries import _SatelliteTimeSeries
@@ -402,260 +397,3 @@ class RestorationArea:
         )
         r80p = r80p.expand_dims(dim={"metric": [str(Metric.R80P)]})
         return r80p
-
-    def plot_spectral_trajectory(self, path: str = None, legend: bool = True) -> None:
-        """Create spectral trajectory plot of the RestorationArea
-
-        Parameters
-        ----------
-        path : str, optional
-            The path to save the plot to.
-        """
-        hist_ref_sys = self.reference_polygons == None
-
-        stats = self.stack.satts.stats()
-        stats = stats.sel(
-            stats=[
-                "median",
-                "mean",
-            ]
-        )
-
-        # convert stats xarray and recovery target xarray into merged df for plotting
-        stats = stats.assign_coords(band=([str(b) for b in stats.band.values]))
-        stats = stats.to_dataframe("value")
-
-        recovery_target = self.recovery_target.assign_coords(
-            band=([str(b) for b in self.recovery_target.band.values])
-        )
-        reco_targets = recovery_target.to_dataframe("reco_targets").dropna(how="any")
-
-        # merge on multi-index: (statistic, band, year) then reset index
-        data = stats.merge(reco_targets, left_index=True, right_index=True)[
-            ["value", "reco_targets"]
-        ]
-        data = data.reset_index()
-        data["time"] = data["time"].apply(lambda x: str(x.year))
-
-        # Set theme and colour palette for plots
-        sns.set_theme()
-        palette = sns.color_palette("deep")
-
-        bands = data["band"].unique()
-        fig, axs = plt.subplots(
-            len(bands), 1, sharey=False, sharex=True, figsize=[8, 7.5]
-        )
-        # Plot per-band statistic lineplots
-        for i, band in enumerate(bands):
-            band_data = data[data["band"] == band]
-            try:
-                axi = axs[i]
-            except TypeError:
-                axi = axs
-
-            sns.lineplot(
-                data=band_data,
-                x="time",
-                hue="stats",
-                y="value",
-                ax=axi,
-                legend=False,
-                lw=1,
-            )
-            sns.lineplot(
-                data=band_data[band_data["stats"] == "mean"],
-                x="time",
-                y="reco_targets",
-                ax=axi,
-                color="black",
-                linestyle=(0, (3, 5, 1, 5)),
-                lw=1,
-            )
-            self._draw_trajectory_windows(axi, palette, hist_ref_sys)
-            self._set_axis_labels(axi, band, data["time"].unique().tolist())
-        (
-            labels,
-            custom_handles,
-        ) = self._custom_legend_labels_handles(palette, hist_ref_sys)
-
-        if legend:
-            plt.figlegend(
-                labels=labels,
-                handles=custom_handles,
-                loc="lower center",
-                fancybox=True,
-                ncol=3,
-                handler_map={Patch: HandlerFilledBetween()},
-            )
-            plt.subplots_adjust(
-                bottom=plt.rcParams["figure.subplot.bottom"]
-                + (plt.rcParams["figure.subplot.bottom"] / 1.5)
-            )
-        if path:
-            plt.savefig(path)
-        else:
-            plt.show()
-
-    def _set_axis_labels(self, axi, title, xlabels):
-        """Set the axis labels to desired values"""
-        axi.set_xticks(
-            axi.get_xticks(),
-            xlabels,
-            rotation=45,
-            ha="right",
-        )
-        axi.set_xlabel("Year")
-        axi.set_ylabel(f"{title} Value")
-
-    def _draw_trajectory_windows(self, axi, palette, hist_ref_sys):
-        """Draw the trajectory windows onto subplots.
-
-        Uses two verticle dashed lines to delimit the start and
-        end years of a window. If the start and end years are
-        not the same year, then the space between the two dashed lines
-        is filled in (vertical span). Each window (i.e line/span group)
-        is coloured a distinct colour.
-
-        Draws the reference, disturbance, and recovery windows.
-
-        """
-        # Draw recovery window
-        axi.axvline(
-            x=self.restoration_start,
-            color=palette[2],
-            linestyle="dashed",
-            lw=1,
-        )
-        axi.axvspan(
-            self.restoration_start,
-            str(self.end_year.year),
-            alpha=0.2,
-            color=palette[2],
-        )
-        axi.axvline(
-            x=str(self.end_year.year),
-            color=palette[2],
-            linestyle="dashed",
-            lw=1,
-        )
-
-        # Draw disturbance window
-        axi.axvline(
-            x=self.disturbance_start,
-            color=palette[3],
-            linestyle="dashed",
-            lw=1,
-        )
-        axi.axvspan(
-            self.disturbance_start,
-            self.restoration_start,
-            alpha=0.2,
-            color=palette[3],
-        )
-
-        if hist_ref_sys:
-            # if deriving target from recovery polygon, draw reference window
-            axi.axvline(
-                x=self.reference_years[0],
-                color=palette[4],
-                linestyle="dashed",
-                lw=1,
-            )
-            axi.axvspan(
-                self.reference_years[0],
-                self.reference_years[1],
-                alpha=0.2,
-                color=palette[4],
-            )
-            # only draw line if reference ye
-            if self.reference_years[1] != self.disturbance_start:
-                axi.axvline(
-                    x=self.reference_years[1],
-                    color=palette[4],
-                    linestyle="dashed",
-                    lw=1,
-                )
-
-    def _custom_legend_labels_handles(self, palette, hist_ref_sys) -> Tuple[List, List]:
-        """Create a custom legend to match trajectory plots
-
-        Returns
-        -------
-        tuple of lists
-            custom labels and handles to pass to ``figlegend``
-
-        """
-        median_line = Line2D([0], [0], color=palette[0], lw=2)
-        mean_line = Line2D([0], [0], color=palette[1], lw=2)
-        recovery_target_line = Line2D(
-            [0], [0], color="black", linestyle=(0, (3, 5, 1, 5)), lw=1
-        )
-        recovery_window_patch = Patch(facecolor=palette[2], alpha=0.2)
-        disturbance_window_patch = Patch(facecolor=palette[3], alpha=0.2)
-        reference_years_patch = Patch(facecolor=palette[4], alpha=0.2)
-
-        custom_handles = [
-            median_line,
-            mean_line,
-            disturbance_window_patch,
-            recovery_window_patch,
-        ]
-
-        labels = [
-            "median",
-            "mean",
-            "disturbance window",
-            "recovery window",
-        ]
-        if hist_ref_sys:
-            if isinstance(self.recovery_target_method, MedianTarget):
-                if self.recovery_target_method.scale == "pixel":
-                    custom_handles.insert(
-                        2,
-                        (recovery_target_line),
-                    )
-                    labels.insert(2, "recovery target (estimated mean)")
-                else:
-                    custom_handles.insert(
-                        2,
-                        recovery_target_line,
-                    )
-                    labels.insert(2, "recovery target")
-
-            custom_handles.insert(3, reference_years_patch)
-            labels.insert(3, "reference year(s)")
-        else:
-            custom_handles.insert(
-                2,
-                recovery_target_line,
-            )
-            labels.insert(2, "recovery target")
-
-        return labels, custom_handles
-
-
-class HandlerFilledBetween(HandlerPatch):
-    """Custom Patch Handler for trajectory windows.
-
-    Draws Patch objects with left and right edges coloured/dashed
-    to match the style of trajectory window Patches in the plots.
-
-    """
-
-    def create_artists(
-        self, legend, orig_handle, xdescent, ydescent, width, height, fontsize, trans
-    ):
-        p = super().create_artists(
-            legend, orig_handle, xdescent, ydescent, width, height, fontsize, trans
-        )[0]
-        color = p.get_facecolor()
-        x0, y0 = 0, 0
-        x1 = x0 + width
-        y1 = y0 + height
-        line_left = Line2D(
-            [x0, x0], [y0, y1], color=color, linestyle="dashed", lw=0.85, alpha=1
-        )
-        line_right = Line2D(
-            [x1, x1], [y0, y1], color=color, linestyle="dashed", lw=0.85, alpha=1
-        )
-        return [p, line_left, line_right]
